feat(ui,backend): add Private Accommodation option for listings
This PR introduces the option to add a listing featuring a private accommodation.
As we had done in the past, choosing a residency will select the location
associated with that residency, and now choosing a private accommodation will
require the user to enter a custom location using the CustomLocationDialog.
For this purpose, I have added a location field to the Rental Listing type.
I have also modified how the listings are fetched in the listings page
under BrowseCityScreen. They are now fetched by location, ensuring that
listings with a custom location are displayed if they belong within the
radius of an available city.
These changes also apply to editing the listings.
I have ensured that the Private Accommodation option does not appear in
the dropdown menu of the residencies in the review as we have decided
that reviews will be for residencies only.

I have also fixed all tests using the RentalListing type, and adapted the
tests of the edit and add listing to ensure correct functionality
in light of the new changes.