refactor(qr): move QR code scanner to `BrowseCityScreen` (#292)
* refactor(qr): move QR code scanner out of settings

Moved the QR code scanner button to `BrowseCityScreen` to the top bar.
It was in the `SettingsScreen` and it didn't really make sense.

* fix(test): update QR code tests

Updated QR code tests previously located in `SettingsScreenTest`.
Moved and adapted those tests in BrowseCityScreenTest.

* fix(qr): QR code button displayed on reviews tab

The QR code scanner button was only displayed on the listings tab. Fixed it
by displaying it also in the reviews tab.